requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.3.0
python-decouple==3.8
pytz==2024.1
drf-spectacular==0.27.2 